+++ src/components/filter/filter-amount/FilterAmount.tsx
@@ -93,7 +93,6 @@ export const FilterAmount = ({
         setNewFilterText(`Less than ${amount1String}`);
         break;
       default:
-        // Handle default case if necessary
         break;
     }
   };
