+++ src/components/filter/filter-date/_DateControls/_IsBetween.tsx
@@ -7,7 +7,7 @@ import {
   DateControlsWrapper,
   defaultDateFormat,
   formatDate,
-} from 'src/components/filter/filter-date/temp';
+} from 'src/components/filter/filter-date/dateUtils';
 import { Input } from 'src/components/input/Input';
 import { Text } from 'src/components/text/Text';
