+++ src/components/filter/filter-date/_DateControls/_IsInTheLast.tsx
@@ -10,7 +10,7 @@ import {
   type _DateControlProps,
   DateControlsWrapper,
   defaultDateFormat,
-} from 'src/components/filter/filter-date/temp';
+} from 'src/components/filter/filter-date/dateUtils';
 import { Input } from 'src/components/input/Input';
 import { Select } from 'src/components/select/Select';
 import type { IOption } from 'src/types/IOption';
